CHANGES: - gradlew permission given. - Moved navigation to independent package. - Moved expectuals to app
