+++ composeApp/src/androidMain/kotlin/org/example/project/app/expectuals/getScreenDimensions.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.expectuals
+package org.example.project.app.expectuals
 
 import android.content.res.Resources
