+++ composeApp/src/commonMain/kotlin/org/example/project/App.kt
@@ -3,7 +3,7 @@ package org.example.project
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.*
 import org.example.project.presentation.theme.AppTheme
-import org.example.project.presentation.navigation.MyAppNavigation
+import org.example.project.navigation.MyAppNavigation
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
