+++ composeApp/src/commonMain/kotlin/org/example/project/app/expectuals/getScreenDimensions.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.expectuals
+package org.example.project.app.expectuals
 
 expect fun getScreenHeight(): Float
 expect fun getScreenWidth(): Float
