+++ composeApp/src/commonMain/kotlin/org/example/project/navigation/Screens.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.navigation
+package org.example.project.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
