+++ composeApp/src/commonMain/kotlin/org/example/project/presentation/profile/ProfilePage.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.currentOrThrow
-import org.example.project.presentation.navigation.Screens
 
 @Composable
 fun ProfilePage() {
