+++ composeApp/src/desktopMain/kotlin/org/example/project/app/expectuals/getViewModelScope.kt
@@ -1,8 +1,9 @@
-package org.example.project.presentation.expectuals
+package org.example.project.app.expectuals
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.swing.Swing
 
 actual fun getViewModelScope(): CoroutineScope =
-    CoroutineScope(SupervisorJob() + Dispatchers.Main)
+    CoroutineScope(SupervisorJob() + Dispatchers.Swing)
